digital-twin/requirements.txt
# Core Dependencies
python-dotenv
pyyaml
click

# Repository Management
gitpython

# Code Parsing
tree-sitter-language-pack
libcst

# Embeddings & Vector Store
sentence-transformers
faiss-cpu
chromadb
langchain
langchain-openai
langchain-huggingface
huggingface_hub
langchain-anthropic
langchain-google-genai

# Search & Retrieval
rank-bm25
networkx

# API
fastapi
uvicorn
pydantic
python-multipart

# Utilities
tqdm
numpy
pandas
pathspec

# Caching
diskcache
redis

# Testing
pytest
pytest-asyncio
pytest-cov


supabase